+++ Projects/Feature/Login/Interface/Sources/GeneralLogIn/GeneralLogInFeatureInterface.swift
@@ -31,3 +31,23 @@ public struct GeneralLogInFeature {
     reducer
   }
 }
+
+extension GeneralLogInFeature {
+  public init() {
+    let reducer = Reduce<State, Action> { state, action in
+      switch action {
+      case .onAppear:
+        return .none
+        
+      case let .presentToastDidRequired(message):
+        // TODO: present toast
+        return .none
+        
+      case let .loginDidCompleted(accessToken, refreshToken):
+        // TODO: Login Completed Handling
+        return .none
+      }
+    }
+    self.init(reducer: reducer)
+  }
+}
